software/requirements.txt
PyQt5>=5.15.0
pyserial
numpy
pyperclip
pyyaml